requirements-web.txt
flask
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv
supabase
